feat(modules): add content publishing and creation workflows
- Introduce `published` status and `parentId` to module types.
- Enable admin actions for publishing, adding new content, and managing
  existing items in the module panel.
- Implement a new route for creating nested module items.
- Update `CMS` component with a `variant` prop to support different
  editor contexts (full-page or embedded).
- Refactor `content-details-editor` to use a reset button, as delete
  functionality is now handled in the module panel.
- Update mock data to reflect new types and filter content by
  publishing status for students.
