--- frontend/src/pages/shared/courses/$courseCode/modules/course-modules.tsx
@@ -1,60 +1,0 @@
-import { Button, Group, Stack, Title } from '@mantine/core'
-import {
-  IconPlus,
-  IconViewportShort,
-  IconViewportTall,
-} from '@tabler/icons-react'
-import React, { useState } from 'react'
-import { useAuth } from '@/features/auth/auth.hook.ts'
-import ModulePanel from '@/features/courses/modules/module-panel.tsx'
-import { Link, useParams } from '@tanstack/react-router'
-
-const CourseModules = () => {
-  const { authUser } = useAuth('protected')
-  const { courseCode } = useParams({ strict: false })
-  const [allExpanded, setAllExpanded] = useState(false)
-
-  const toggleExpandAll = () => {
-    setAllExpanded((prev) => !prev)
-  }
-
-  return (
-    <Stack gap={'md'} p={'md'}>
-      {/*Header*/}
-      <Group justify="space-between" align="center">
-        <Title>Modules</Title>
-        <Group align="center">
-          <Button
-            onClick={toggleExpandAll}
-            w={135}
-            variant={'default'}
-            leftSection={
-              allExpanded ? (
-                <IconViewportShort size={16} />
-              ) : (
-                <IconViewportTall size={16} />
-              )
-            }
-          >
-            {allExpanded ? 'Collapse All' : 'Expand All'}
-          </Button>
-          {authUser.role === 'admin' && (
-            <>
-              <Button
-                leftSection={<IconPlus />}
-                bg={'secondary'}
-                component={Link}
-                to={`./create`}
-              >
-                Add New Content
-              </Button>
-            </>
-          )}
-        </Group>
-      </Group>
-      <ModulePanel allExpanded={allExpanded} />
-    </Stack>
-  )
-}
-
-export default CourseModules
